ENH: Adds graph convolution models

+++ fcnn/models/pytorch/graph/utils.py
@@ -22,8 +22,8 @@ def faces_to_edges(faces):
 
 def faces_to_adjacency_matrix(faces, size):
     indices = faces_to_edges(faces)
-    values = torch.zeros(indices.shape[1], dtype=torch.long)
-    adjacency_matrix = torch.sparse.LongTensor(indices, values, size=size)
+    values = torch.zeros(indices.shape[1], dtype=torch.float)
+    adjacency_matrix = torch.sparse.FloatTensor(indices, values, size=size)
     return adjacency_matrix
